Update defaults for high resolution files.
The defaults for 7.5km and 3.75km files have been updated.
An entry has been added for a 60-3km regionally refined mesh,
centered over the western U.S. Additionally, a hack has been added
to permit the specification of a restart file to start the code
for a hybrid or branch run.